// src/components/Diary/LineArrow.d.ts
/// <reference types="react" />
export declare const Line: ({ height, width, className }: {
    height?: number | undefined;
    width?: number | undefined;
    className?: string | undefined;
}) => JSX.Element;
export declare const ArrowDown: ({ height, width, className }: {
    height?: number | undefined;
    width?: number | undefined;
    className?: string | undefined;
}) => JSX.Element;